till product details

+++ lib/domain/models/inventory/get_inventory_response_model/datum.dart
@@ -12,13 +12,13 @@ class Inventory {
   String? productName;
   String? size;
   int? stock;
-  int? price;
+  double? price;
   @JsonKey(name: 'if_present_at_wishlist')
   bool? ifPresentAtWishlist;
   @JsonKey(name: 'if_present_at_cart')
   bool? ifPresentAtCart;
   @JsonKey(name: 'discounted_price')
-  int? discountedPrice;
+  double? discountedPrice;
 
   Inventory({
     this.id,
